% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de-circ.R
\name{de_circrna}
\alias{de_circrna}
\title{EasyCircR - DE circRNAs}
\usage{
de_circrna(
  circ_mtx,
  condition,
  design,
  contr,
  min_num_samples = 2,
  lfc = 1,
  p.value = 0.05,
  voomWithQualityWeights = FALSE,
  plotVoomResult = FALSE,
  ...
)
}
\arguments{
\item{circ_mtx}{the count matrix of circRNAs as results of \code{EasyCircR::read_ciri_output()}.}

\item{condition}{\code{numeric}, \code{character} or \code{factor} with two levels
that groups the samples into two conditions.}

\item{design}{design matrix with rows corresponding to samples and columns to coefficients
to be estimated.}

\item{contr}{contrast matrix computed by \code{limma::makeContrasts()}.}

\item{min_num_samples}{removing all the circRNAs that are not present in at least the specified number of samples. Default is \code{2}.}

\item{lfc}{minimum absolute log2-fold-change required.}

\item{p.value}{Benjamini-Hochberg adjusted p-values cutoff. Only genes with lower p-values are listed. Default is \code{0.05}.}

\item{voomWithQualityWeights}{\code{logical}. If \code{TRUE} combine voom observational-level weights with sample-specific quality weights in a designed experiment.}

\item{plotVoomResult}{\code{logical}. If \code{TRUE} the plot of the mean-variance trend and sample-specific weights is displayed.}

\item{...}{other arguments are passed to \code{limma:topTable()}.}
}
\value{
A \code{dataframe} storing for each circRNA following features: \describe{
\item{logFC}{estimate of the log2-fold-change corresponding to the effect or contrast}
\item{AveExpr}{average log2-expression of circRNA in all the samples}
\item{t}{moderated t-statistic}
\item{P.Value}{raw p-value}
\item{adj.P.Value}{adjusted Benjamini-Hochberg p-value}
\item{B}{log-odds that the circRNA is differential expressed}}
}
\description{
Differential expressed analysis of detected circRNAs performed with \code{limma} package.
}
\examples{
design <- model.matrix(~0+condition)
contr <- limma::makeContrasts("PQRvsDMSO"  = conditionPQR - conditionDMSO, levels = colnames(design))

circ_de <- de_circrna(circ_mtx, condition, design, contr, lfc=0, p.value=0.05, 
                      voomWithQualityWeights=FALSE, min_num_samples=2)

}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
