% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirna-binding.R
\name{get_mirna_binding}
\alias{get_mirna_binding}
\title{EasyCircR - miRNA response element prediction}
\usage{
get_mirna_binding(circ_df, force = FALSE)
}
\arguments{
\item{circ_df}{the \code{data.frame} containing circRNAs features as results of \code{EasyCircR::read_ciri_output()}.}

\item{force}{\code{logical(1)}. If \code{FALSE} the tool will search for the already stored RDS output files, if
there are none it will generate them. If \code{TRUE} it will force the redo of all the steps of the function.}
}
\value{
a \code{data.frame} storing associations between each circRNA and one or more miRNA families.
}
\description{
Predict miRNA response elements (MREs) from the full sequence of circRNA using TargetScan.
Check if \code(circ_mirna.txt) is stored in "EasyCirc/circRNA/miRNA" otherwise execute the step (see \code{force} parameter).
}
\examples{
# read CIRI output
circ <- read_ciri_output(samples_file)
names(circ)
circ_df <- circ$circ_df
circ_mtx <- circ$circ_mtx

# identify DE circRNAs
design <- model.matrix(...) #define your model 
contr <- limma::makeContrasts(...) #define your contrasts 
circ_de <- de_circrna(...) #define your parameters 

#Select only DE circRNAs
circ_df_de <- circ_df[circ_df$bsj_id \%in\% rownames(circ_de), ]


circ_mirna <- get_mirna_binding(circ_df_de, force=FALSE)

}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
