% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-regulated-gene.R
\name{get_postregulated_gene}
\alias{get_postregulated_gene}
\title{EasyCircR - Quantify post-transcriptional regulation of gene expression}
\usage{
get_postregulated_gene(
  samples_file,
  genome_file,
  genome_annotation_file,
  condition,
  outdir = ".",
  aligner = "Rhisat2",
  force = FALSE,
  stranded_data = TRUE,
  n_core = 1,
  cacheDir = NULL
)
}
\arguments{
\item{samples_file}{path to tab separated file with three named columns indicating:
two pair-end fastq files and sample name.  Column names are FileName1,
FileName2 and SampleName. A number of samples greater than 2 is required for each condition.}

\item{genome_file}{path to genome file.}

\item{genome_annotation_file}{path to annotation file.}

\item{condition}{\code{numeric}, \code{character} or \code{factor} with two levels
that groups the samples into two conditions. The contrast will be defined as secondLevel - firstLevel.}

\item{outdir}{path to the output directory 'EasyCirc' that store the final results. The three major folders ("circRNA",
"geneMirnaCirc", "postGene") contain each one or two \code{.rds} file storing the results of the pipeline,
e.g. \code{postGene.rds} the results of \code{get_postregulated_gene} or \code{countMatrix.rds} contains the counts of circRNA found.}

\item{aligner}{selects the aligner program to be used for aligning the reads for the EISA step. Currently, only "Rbowtie" and
"Rhisat2" (default) are supported.}

\item{force}{\code{logical(1)}. If \code{FALSE} the tool will search for the already stored RDS output files, if
there are none it will generate them. If \code{TRUE} it will force the redo of all the steps of the function.}

\item{stranded_data}{\code{logical(1)}. If \code{TRUE}, the RNA-seq data is assumed to be strand-specific, and therefore only overlapping genes that are on the same strand will be filtered out during the EISA analysis If \code{FALSE}, also genes overlapping on opposite strands will be filtered out.}

\item{n_core}{number of cores to be used. Default is \code{1}.}

\item{cacheDir}{specifies the location to store (potentially huge) aligner temporary files . If set to \code{NULL} (default), the temporary directory of the current R session as returned by \code{tempdir()} will be used.}
}
\value{
a \code{data.frame} with elements that stores statisical results for differential changes
between exonic and intronic contrast, an indication for post-transcriptional regulation.
}
\description{
Measure changes across different experimental conditions to quantify post-transcriptional regulation of gene expression (EISA).
}
\examples{
samples_file <- "samples.txt"
genome_file <- "genome/hg38.fa"
genome_annotation_file <- "genome/hg38.gtf"
condition <- factor(rep(c("DMSO","PQR"),2))

post_gene <- get_postregulated_gene(samples_file, genome_file, genome_annotation_file, 
                                   condition, aligner="Rbowtie",
                                   force=FALSE, n_core=10)
                                   
post_gene <- post_gene[post_gene$FDR <= 0.1,] 
                                                                       
}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
