% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulatory-network.R
\name{plot_regulatory_net}
\alias{plot_regulatory_net}
\title{EasyCircR - plot the regulatory network}
\usage{
plot_regulatory_net(bsj_id, gene_mirna_circ = NULL)
}
\arguments{
\item{bsj_id}{the Back-Splice junction ID of the circRNA user wants to display.}

\item{gene_mirna_circ}{the \code{data.frame} as results of \code{EasyCircR::connect_circ_gene(...)}.
By default (\code{NULL}) the function check if the \code{geneMirnaCirc.rds} is stored in "EasyCirc/geneMirnaCirc",
otherwise execute \code{EasyCircR::connect_circ_gene()}.}
}
\description{
Plot the circRNA-miRNA-gene regulatory network.
}
\examples{
plot_regulatory_net("12:116230533|116237705:-:13")

}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
