% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri-output.R
\name{read_ciri_output}
\alias{read_ciri_output}
\title{EasyCircR - read CIRI-full output}
\usage{
read_ciri_output(sampleFile = NULL, chrMT = T)
}
\arguments{
\item{sampleFile}{path to tab separated file with three named columns indicating:
two pair-end fastq files and sample name.  Column names are FileName1,
FileName2 and SampleName. A number of samples greater than 2 is required for each condition.}

\item{chrMT}{\code{logical}. \code{FALSE} if the user wants to exclude circRNAs detected on MT chromosome.}
}
\value{
a \code{list} with two elements: \describe{
\item{circ_df}{a \code{dataframe} containing information like the chromosome, the start and end of the back-spliced junction (BSJ),
the length of the reconstructed circRNA, the reconstructed sequence and the circBank annotation.}
\item{circ_mtx}{is the count matrix of circRNAs.}}
}
\description{
Read CIRI-full output starting from the count matrix and description file stored in the "EasyCirc/circRNA/CIRI-full" directory.
Additionally, the function maps significant differentially expressed circRNAs to their host gene annotation supplied by the circBank database.
}
\examples{
circ <- read_ciri_output(samples_file)
names(circ)
circ_df <- circ$circ_df
circ_mtx <- circ$circ_mtx
  
}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
