% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri-full.R
\name{run_ciri_full}
\alias{run_ciri_full}
\title{EasyCircR - detection of circRNAs}
\usage{
run_ciri_full(
  samples_file,
  genome_file,
  genome_annotation_file,
  trim_reads_length = NULL,
  force = FALSE,
  n_core = parallel::detectCores(),
  remove_temporary_files = TRUE
)
}
\arguments{
\item{samples_file}{path to tab separated file with three named columns indicating:
two pair-end fastq files and sample name.  Column names are FileName1,
FileName2 and SampleName. A number of samples greater than 2 is required for each condition.}

\item{genome_file}{path to genome file.}

\item{genome_annotation_file}{path to annotation file.}

\item{trim_reads_length}{\code{numeric}, define at which length reads must be cutted.
Reads smaller than that size are discarded and longer are trimmed.
CIRI-full cannot process sequencing reads with different lengths.
Trimming results are stored in "EasyCirc/circRNA/trimmed_fastq" directory.}

\item{force}{\code{logical(1)}. If \code{FALSE} the tool will search for the already stored RDS output files, if
there are none it will generate them. If \code{TRUE} it will force the redo of all the steps of the function.}

\item{n_core}{\code{numeric}, the number of CPU cores to use. If not specified, EasyCircR
tries to detect the number of CPU cores on the current host using the function \code{parallel:detectCores()}.}

\item{remove_temporary_files}{\code{logical}, if \code{TRUE} temporary files will be removed at the end of each sample circRNA detection step.
EasyCircR requires a significant amount of disk memory, it is recommended to leave the default value \code{TRUE}}
}
\value{
the function generates an output directory called "CIRI-Full" (in "EasyCirc" working directory) in which are stored all the CIRI-full analysis
results including \describe{
\item{countMatrix.rds}{count matrix of circRNAs}
\item{predictedCircRNAs.rds}{contains information like the chromosome, the start and end of the back-spliced junction (BSJ),
the length of the reconstructed circRNA and the reconstructed sequence.}
\item{sample directory}{a directory for each sample where all the temporary and final CIRI-full results are stored.}
}
}
\description{
Reconstruction and quantification full-length circRNAs with CIRI-full.
}
\examples{
samples_file <- "samples.txt"
genome_file <- "genome/hg38.fa"
genome_annotation_file <- "genome/hg38.gtf"
run_ciri_full(samples_file, genome_file, genome_annotation_file, 
              force=FALSE, trim_reads_length=130, n_core= 10)
              
}
\author{
Luca Parmigiani, Antonino Aparo, Simone Avesani
}
