% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-easycirc.R
\name{run_easycirc}
\alias{run_easycirc}
\title{Run EasyCirc Pipeline}
\usage{
run_easycirc(
  samples_file,
  genome_file,
  genome_annotation_file,
  condition,
  force = FALSE,
  eisa_n_core = 1,
  cirifull_n_core = 1,
  percent_of_reads_after_trim = 0.8,
  lfc = 1,
  p.value = 0.05,
  mirna_gene_type = "validated",
  predicted.cutoff = 35,
  eisa_aligner = "Rhisat2",
  eisa_stranded_data = TRUE,
  circrna_filter_min_num_samples = 2,
  voomWithQualityWeights = FALSE,
  remove_temporary_files = TRUE,
  trim_reads_length = NULL
)
}
\arguments{
\item{samples_file}{path to tab separated file with three named columns indicating:
two pair-end fastq files and sample name.  Column names are FileName1,
FileName2 and SampleName.}

\item{genome_file}{path to genome file.}

\item{genome_annotation_file}{path to annotation file.}

\item{condition}{\code{numeric}, \code{character} or \code{factor} with two levels
that groups the samples (columns of \code{cntEx} and \code{cntIn}) into two
conditions. The contrast will be defined as secondLevel - firstLevel.}

\item{force}{\code{logical(1)}. If \code{FALSE} it will search for the already stored RDS output files, if
there are none it will generate them. If \code{TRUE} it will force the redo of all
the steps of the pipeline.}

\item{eisa_n_core}{number of core to be used in the EISA step. Default is \code{1}.}

\item{cirifull_n_core}{number of core to be used in the CIRI-Full step. Default is \code{1}.}

\item{percent_of_reads_after_trim}{Percentage of reads to be kept after the trimming of the #' reads. Default is \code{.8}.
\describe{Since CIRI-Full relies on CIRI-AS it
cannot process sequencing reads with different lengths. The function
\code{run_easycirc} accept the parameter \code{trim_reads_length} which
define at which length reads must be cutted. Reads smaller than that size
are discarded and longer are trimmed. In order to keep most of the reads
and as long as possible, the function \code{plot_readlen_hist} can help. In
the dataset all the reads have already been trimmed and the function returns
the length of the reads. \code{trim_reads_length} is automatically computed from
\code{percent_of_reads_after_trim}}}

\item{lfc}{Minimum absolute log2-fold-change required for circRNA and EISA differential
expressed step.}

\item{p.value}{cutoff value for adjusted p-values. Only genes/circRNAs with lower
p-values are listed.}

\item{mirna_gene_type}{used in the prediction between miRNA and gene interaction.
A character string indicating which table(s) in multiMiR to search.  Each
table contains data from an external database.  Options include "validated"
(default, to search all validated tables "mirecords", "mirtarbase", and
"tarbase"), "predicted" (to search all predicted tables "diana_microt",
"elmmo", "microcosm", "miranda", "mirdb", "pictar", "pita", and
"targetscan"), "disease.drug" (to search all disease/drug tables
"mir2disease", "pharmaco_mir", and "phenomir"), "all" (to search all of the
tables above), or an individual table from above}

\item{predicted.cutoff}{'NULL' or an integer giving a prediction
score cutoff.  By default the cutoff is '32'.}

\item{eisa_aligner}{selects the aligner program to be used for aligning the
reads for the eisa step. Currently, only “Rbowtie” and “Rhisat2” (default) are
supported.}

\item{eisa_stranded_data}{\code{logical(1)}. If \code{TRUE}, the RNA-seq
data is assumed to be strand-specific, and therefore only overlapping genes
that are on the same strand will be filtered out during the eisa step. If
\code{FALSE}, also genes overlapping on opposite strands will be filtered
out.}

\item{circrna_filter_min_num_samples}{\code{numeric}. Before computing the
DE analysis on circRNA filter out circRNAs that are not expressed in at
least \code{circrna_filter_min_num_samples} samples. Default is \code{2}.}

\item{voomWithQualityWeights}{\code{logical(1)}. Use
\code{voomWithQualityWeights} instead of \code{voom} during circRNA DE
analisis.}

\item{remove_temporary_files}{\code{logical(1)}. If \code{TRUE} (Default) alignment
and extra files generated during the CIRI-Full step are deleted after
completion. CIRI-Full can store more than 15GB of disk for each sample,
starting from pair-end sequencing with ~2GB size.}
}
\value{
a \code{list} with elements \describe{
\item{circ_de}{statisical results for differential changes between circRNAs
expression in the two different conditions.}
\item{gene_de}{statisical results for differential changes between exonic
and intronic contrast, an indication for post-transcriptional regulation.}
\item{circ_df}{data.frame containing circRNA information, such as the
start and end of the back-spliced junction (BSJ), the length of the
reconstructed circRNA and the reconstructed sequence.}
\item{gene_mirna_circ}{Predicted relations circRNA-miRNA with known/predicted miRNA-gene.}
}
}
\description{
Starting from RNA-seq data, it uses CIRI-full for the reconstruction of
full-length circRNAs and EISA for the detection of genes which are
post-transcriptional regulated. In order to combine the two results, miRNA
response elements (MREs), with respective binding miRNA, are predicted using
TargetScan for each circRNA.  The relationship miRNA-gene is retrieved using
MultimiR package and results are then filtered by genes which where previously
found post-transcriptional regulated.

The pipeline is divided in 5 major steps.
\enumerate{
\item Measure changes across different experimental conditions to quantify post-transcriptional regulation of gene expression (EISA).
\item Reconstruct and quantify full-length circRNAs with CIRI-Full.
\item Differentially expressed analysis of detected circRNAs (with limma).
\item Predict MREs from the full sequence of circRNA (TargetScan, miRanda, RNA-hybrid).
\item Finally, connecting circRNA to genes.
}
}
\examples{
samples_file <- "test/samples_test.txt"  
genome_file <- "test/genome/hg38_chr3_chr7.fa"  
genome_annotation_file <- "test/genome/hg38_chr3_chr7.gtf"  
condition <- factor(rep(c("DMSO", "PQR"),3))  

res <- run_easycirc(samples_file, genome_file, genome_annotation_file, condition)  
res$circ_de  
res$gene_de  
head(res$circ_df)  
head(res$gene_mirna_circ)  

}
\author{
Luca Parmigiani
}
